renamed get/start next task to claim next task per  PR feedback
